should find all hero test
